+++ Model/Items/Toy.cs
@@ -8,7 +8,7 @@ namespace Ultimate_Tamagochi.Model.Items
     public class Toy : Item
     {
         public int Epicness { get; set; }
-        public Toy(string name, float spawnChance, bool isConsumible, int epicness)
+        public Toy(string name, int spawnChance, bool isConsumible, int epicness)
         {
             Name = name;
             SpawnChance = spawnChance;
